feat: enhance subscription page for management
